Refactor Project structure
